src/mcp-server/requirements.txt
mcp[http]>=1.0.0
azure-identity>=1.15.0
aiohttp>=3.9.0
httpx>=0.27.0
PyJWT[crypto]>=2.8.0
cryptography>=42.0.0
uvicorn>=0.30.0
starlette>=0.38.0
azure-storage-blob[aio]>=12.19.0
